Add read methods to convert from afw formats.
Unfortunately the test failures here mean that our original plan for
rewriting afw images will not work, because re-quantizing with the
same parameters *does* change the values.  There are some comments in
the CFITSIO source code that suggest that they were trying to make that
more likely, but no other indication it should even be possible, and
it's a tricky problem that might be impossible.